added several methods for working with problems,course-problems

+++ src/main/java/com/mit/user/userservice/controller/CourseProblemsController.java
@@ -1,26 +1,25 @@
 package com.mit.user.userservice.controller;
 
+import com.mit.user.userservice.model.Course;
+import com.mit.user.userservice.model.CoursesRepository;
 import com.mit.user.userservice.model.Problem;
 import com.mit.user.userservice.model.ProblemRepository;
-import com.mit.user.userservice.model.User;
-import com.mit.user.userservice.model.UserDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.ModelAndView;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 @RestController
 @RequestMapping("/courses/{courseId}/problems")
 public class CourseProblemsController {
     private final ProblemRepository problemRepository;
+    private final CoursesRepository coursesRepository;
 
-    public CourseProblemsController(ProblemRepository problemRepository) {
+    public CourseProblemsController(ProblemRepository problemRepository, CoursesRepository coursesRepository) {
         this.problemRepository = problemRepository;
+        this.coursesRepository = coursesRepository;
     }
 
     @GetMapping(path = "/", produces = "application/json")
@@ -38,30 +37,33 @@ public class CourseProblemsController {
     }
 
     @PostMapping(path = "/", consumes = "application/json")
-    public void addProblemToCourse(@PathVariable long courseId,
-                                   @RequestParam(name = "problemId") long problemId) {
-        problemRepository.addProblem(courseId, problemId);
+    public ResponseEntity addProblemToCourse(@PathVariable long courseId,
+                                             @RequestParam(name = "problemId") long problemId) {
+        Optional<Problem> problem = problemRepository.findById(problemId);
+        if (problem.isPresent()) {
+            Optional<Course> course = coursesRepository.getCourseById(courseId);
+            if (course.isPresent()) {
+                Optional<Problem> existsProblem = problemRepository.getProblemById(courseId,problemId);
+                if (!existsProblem.isPresent()) {
+                    problemRepository.addProblem(courseId, problemId);
+                    return new ResponseEntity("Problem was successfully added", HttpStatus.OK);
+                } else {
+                    return new ResponseEntity("We already have this problem on this course", HttpStatus.BAD_REQUEST);
+                }
+            }
+            return new ResponseEntity<>("Course with this id was not found", HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>("Problem with this id was not found", HttpStatus.NOT_FOUND);
     }
-//
-//    @GetMapping(path = "{problemId}", produces = "application/json")
-//    public Optional<Problem> getProblemById(@PathVariable long problemId) {
-//        return problemRepository.findById(problemId);
-//    }
-//
-//    @PutMapping(path = "{problemId}")
-//    public void updateProblem(@PathVariable long problemId, @RequestBody Problem problem) {
-//        Long id = problem.getId();
-//        problemRepository.updateProblem(problem.getId(), problem.getProblemName(), problem.getProblemText(), problem.getProblemTime());
-//    }
-//
-//    @DeleteMapping(path = "{problemId}")
-//    public void deleteProblem(@PathVariable long problemId) {
-//        problemRepository.deleteById(problemId);
-//    }
 
     @DeleteMapping(path = "/")
-    public void deleteProblemFromCourse(@PathVariable long courseId,
-                                        @RequestParam(name = "problemId") long problemId) {
-        problemRepository.deleteProblemFromCourse(courseId, problemId);
+    public ResponseEntity deleteProblemFromCourse(@PathVariable long courseId,
+                                                  @RequestParam(name = "problemId") long problemId) {
+        Optional<Problem> problemTemp = problemRepository.findById(problemId);
+        if (problemTemp.isPresent()) {
+            problemRepository.deleteProblemFromCourse(courseId, problemId);
+            return new ResponseEntity<>("Problem was successfully deleted", HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Problem with this id was not found", HttpStatus.BAD_REQUEST);
     }
 }
